update vap sample comp and pipeline#6
